### 2020-01-17 #2022 @kyleniemeyer
@jspaezp FYI, it looks like your article has an invalid DOI for one of the referenes
