### 2020-01-21 #2022 @natallah
Patro, R., Duggal, G., Love, M. et al. Salmon provides fast and bias-aware quantification of transcript expression. Nat Methods 14, 417–419 (2017) doi:10.1038/nmeth.4197

The above is the issue.  Sebastian is aware and will comment soon
