### 2020-01-21 #2022 @kyleniemeyer
@natallah  the DOI 10.1038/nmeth.4197.Salmon does not resolve (try https://doi.org/10.1038/nmeth.4197.Salmon), it leads to an error
