### 2020-01-22 3385a8f
fixed DOI in references
