+++ vignettes/paper/bibliography.bib
@@ -70,7 +70,7 @@ year = {2018}
 @article{Patro2017,
 abstract = {We introduce Salmon, a method for quantifying transcript abundance from RNA-seq reads that is accurate and fast. Salmon is the first transcriptome-wide quantifier to correct for fragment GC content bias, which we demonstrate substantially improves the accuracy of abundance estimates and the reliability of subsequent differential expression analysis. Salmon combines a new dual-phase parallel inference algorithm and feature-rich bias models with an ultra-fast read mapping procedure. },
 author = {Patro, Rob and Duggal, Geet and Love, Michael I and Irizarry, Rafael A and Kingsford, Carl},
-doi = {10.1038/nmeth.4197.Salmon},
+doi = {10.1038/nmeth.4197},
 journal = {Nature Methods},
 mendeley-groups = {Sequencing},
 number = {4},
