### 2020-01-22 #2022 @jspaezp
Thanks @kyleniemeyer and @natallah, I commited and pushed the modified DOI.
