### 2020-01-23 #2022 @will-rowe
Hi @jspaezp - sorry for the delay here. Do you have any reviewer suggestions? If not, I will consult our JOSS list to see if I can identify some suitable people.
