### 2020-01-23 #2022 @kyleniemeyer
@natallah thanks for the suggestions, but those are from our editor list—could you look at the [reviewer list][(https://bit.ly/joss-reviewers) instead?
