### 2020-01-24 #2022 @will-rowe
Hi @afrubin and  @mschubert - would you please be able to review this submission for JOSS?
