### 2020-01-27 #2022 @will-rowe
Thanks @mschubert! I don't seem to have edit rights on the JOSS reviewers doc - pinging @arfon to see if he can let us know how to update reviewer topics
