### 2020-01-27 #2022 @arfon
> Side note @will-rowe: is there a way to update my topics on the reviewer list? (there's many more reviewers than when I put them in, so they should be more detailed now)

@mschubert - could you email me directly on arfon.smith@gmail.com to request these changes?
