### 2020-02-03 #2061 @whedon
[REVIEW]: scTree: An R package to generate antibody-compatible classifiers from single-cell sequencing data
**Submitting author:** @jspaezp (<a href="http://orcid.org/0000-0002-0065-1474">J. Sebastian Paez</a>)
**Repository:** <a href="https://github.com/jspaezp/sctree" target ="_blank">https://github.com/jspaezp/sctree</a>
**Version:** 0.0.5
