### 2020-02-03 jspaezp/sctree#2 @mschubert
[joss review] community guidelines
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2061).

The JOSS checklist states:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Please provide this information (usually done in the `README` or `CONTRIBUTING.md`)
