### 2020-02-03 jspaezp/sctree#3 @mschubert
[joss review] automated tests
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2061).

Can you please make sure your package passes [its tests](https://travis-ci.org/jspaezp/sctree)?

Currently, there is a URL error on Travis for vignette building.
