### 2020-02-03 4bbc682
Create Contribute.md
Created file to detail guidelines for contributions, error/issue reporting, and contact
