+++ Contribute.md
@@ -1,0 +1,8 @@
+# Contributing
+We welcome contributions, which will take place on Github.  Feel free to suggest edits to code, examples, vignete, and documentation!  We are happy to hear your suggestions and likewise happy for you contributions! Soure code is here https://github.com/jspaezp/sctree/tree/master/
+
+# Report Issues
+Please report issues here: https://github.com/jspaezp/sctree/issues
+
+# Seek Support
+If you have any issues please contact natallah@purdue.edu
