### 2020-02-03 jspaezp/sctree#2 @natallah
I added Contribute.md

Thank you!
