### 2020-02-04 jspaezp/sctree#2 @jspaezp
I will mark the issue as closed, and say that it was addressed in 4bbc6825f45f3738c63618592b9b14393c9c755c

Thank you very much for the note and thanks @natallah for adding the document
Let me know if any additional specifications are required.
