### 2020-02-04 a674d18
Added assertion to skip testing that fails in travis, badge to readme and unified antibody querying documentation
