+++ man/query_cc_antibodies.Rd
@@ -1,45 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/antibodies.R
-\name{query_cc_antibodies}
-\alias{query_cc_antibodies}
-\title{Query cell signaling for Single cell antibodies}
-\usage{
-query_cc_antibodies(search_term)
-}
-\arguments{
-\item{search_term}{The term used to query the cell signaling webpage}
-}
-\value{
-NULL when no results are found or a DF with information
-}
-\description{
-Queries Cell Signaling Technologies and
-filters for antibodies for IFC, IHC, and FLOW.
-}
-\details{
-Note that it is just a wrapper arround the web search of the site, so please
-be nice on them and dont get your IP banned ...
-}
-\examples{
-query_cc_antibodies("CD11bfakename")
-# NULL
-query_cc_antibodies("CD11c")
-#   Cat_num                                       Product Name Reactivity
-# 1   97585                           CD11c (D1V9Y) Rabbit mAb          M
-# 2   97473 CD11c (3.9) Mouse mAb (violetFluor™ 450 Conjugate)          H
-# 3   42847    CD11c (3.9) Mouse mAb (redFluor™ 710 Conjugate)          H
-# 4   80342          CD11c (3.9) Mouse mAb (PE-Cy7® Conjugate)          H
-# 5   56025               CD11c (3.9) Mouse mAb (PE Conjugate)          H
-# 6   69627             CD11c (3.9) Mouse mAb (FITC Conjugate)          H
-# 7   51562         CD11c (3.9) Mouse mAb (APC-Cy7® Conjugate)          H
-# 8   36268              CD11c (3.9) Mouse mAb (APC Conjugate)          H
-# Product_Type
-# 1  Primary Antibodies
-# 2 Antibody Conjugates
-# 3 Antibody Conjugates
-# 4 Antibody Conjugates
-# 5 Antibody Conjugates
-# 6 Antibody Conjugates
-# 7 Antibody Conjugates
-# 8 Antibody Conjugates
-}
