+++ man/query_biolegend_antibodies.Rd
@@ -1,33 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/antibodies.R
-\name{query_biolegend_antibodies}
-\alias{query_biolegend_antibodies}
-\title{Query Biolegend for antibodies}
-\usage{
-query_biolegend_antibodies(search_term)
-}
-\arguments{
-\item{search_term}{The term used to query the webpage}
-}
-\value{
-a length 0 character when no results are found or a character vector with the product names
-}
-\description{
-Queries Biolegend for antibodies and
-filters for antibodies for Flow Cytometry
-}
-\details{
-Note that it is just a wrapper arround the web search of the site, so please
-be nice on them and dont get your IP banned ...
-}
-\examples{
-query_biolegend_antibodies("CD11bfakename")
-# NULL
-head(query_biolegend_antibodies("CD11C"))
-# [1] "MojoSortÃ¢\\u0084Â¢ Mouse CD11c Nanobeads"
-# [2] "APC anti-human CD11c Antibody"
-# [3] "Biotin anti-human CD11c Antibody"
-# [4] "FITC anti-human CD11c Antibody"
-# [5] "PE anti-human CD11c Antibody"
-# [6] "PE/Cy5 anti-human CD11c Antibody"
-}
