### 2020-02-04 6e5b875
modified example testing adding exeptions for travis and changed the api to comply with santa cruz
