### 2020-02-04 e69c96b
fixed additional case for the sc api
