+++ R/seurat_utils.R
@@ -53,7 +53,7 @@ NULL
 #'
 #' # Take all cells in cluster 2, and find markers that separate cells in the 'g1' group (metadata
 #' # variable 'group')
-#' markers <- FindMarkers(pbmc_small, ident.1 = "g1", group.by = 'groups', subset.ident = "2")
+#' suppressWarnings(markers <- FindMarkers(pbmc_small, ident.1 = "g1", group.by = 'groups', subset.ident = "2"))
 #' head(x = markers)
 #'
 #' # Pass 'clustertree' or an object of class phylo to ident.1 and
@@ -103,7 +103,7 @@ FindMarkers <- function(object, ...) {
 #'
 #' @examples
 #' # Find markers for all clusters
-#' all.markers <- FindAllMarkers(object = pbmc_small)
+#' suppressWarnings(all.markers <- FindAllMarkers(object = pbmc_small))
 #' head(x = all.markers)
 #' \dontrun{
 #' # Pass a value to node as a replacement for FindAllMarkersNode
