+++ R/antibodies.R
@@ -56,12 +56,12 @@
 #' # NULL
 #' head(query_sc_antibodies("CD11C"))
 #' # Product_Name  Cat_num Citations Rating     Epitope
-#' # 1   Integrin alpha X Antibody (B-6) sc-46676         3    (8) Integrin αX
-#' # 2   Integrin alpha X Antibody (3.9)  sc-1185         2    (2) Integrin αX
-#' # 3 Integrin alpha X Antibody (B-Iy6) sc-19989         1    (1) Integrin αX
-#' # 4  Integrin alpha X Antibody (N418) sc-23951         3    (2) Integrin αX
-#' # 6 Integrin alpha X Antibody (2Q865) sc-71455       NEW    (1) Integrin αX
-#' # 7 Integrin alpha X Antibody (3H986) sc-71456       NEW    (1) Integrin αX
+#' # 1   Integrin alpha X Antibody (B-6) sc-46676         3    (8) Integrin aX
+#' # 2   Integrin alpha X Antibody (3.9)  sc-1185         2    (2) Integrin aX
+#' # 3 Integrin alpha X Antibody (B-Iy6) sc-19989         1    (1) Integrin aX
+#' # 4  Integrin alpha X Antibody (N418) sc-23951         3    (2) Integrin aX
+#' # 6 Integrin alpha X Antibody (2Q865) sc-71455       NEW    (1) Integrin aX
+#' # 7 Integrin alpha X Antibody (3H986) sc-71456       NEW    (1) Integrin aX
 #' # Species                       Method
 #' # 1   human WB, IP, IF, IHC(P) and ELISA
 #' # 2   human               IP, IF and FCM
@@ -130,7 +130,6 @@ query_cc_antibodies <- function(search_term, sleep = 1) {
 }
 
 
-
 #' @rdname antibodies
 #' @export
 query_sc_antibodies <- function(search_term, sleep = 1) {
@@ -140,16 +139,30 @@ query_sc_antibodies <- function(search_term, sleep = 1) {
         "No=0&Nrpp=30&Ntt=", search_term)
 
     url <- url(url, "rb")
-    antibody_df <- rvest::html_nodes(
-        xml2::read_html(url),
-        xpath="/html/body/div[1]/div/div/div/div[4]/table")
+    corpus <- xml2::read_html(url)
     close(url)
 
-    antibody_df <- rvest::html_table(antibody_df, fill = TRUE)
+    antibody_df <- rvest::html_table(
+        rvest::html_nodes(
+            corpus,
+            xpath="/html/body/div[1]/div/div/div/div[4]/table"),
+        fill = TRUE)
+
+    backup_antibody_df <- rvest::html_table(
+        rvest::html_nodes(
+            corpus,
+            xpath="/html/body/div[1]/div/div/div/div[5]/table"),
+        fill = TRUE)
+
     # The former will be a list of length 0 if no results are found
 
-    if (length(antibody_df) == 0) {return(NULL)}
-
+    if (length(antibody_df) == 0) {
+        if (length(backup_antibody_df) == 0) {
+            return(NULL)
+        } else {
+            antibody_df <- backup_antibody_df
+        }
+    }
 
     antibody_df <- antibody_df[[1]]
     antibody_df <- antibody_df[,2:(ncol(antibody_df) - 1)]
