+++ man/antibodies.Rd
@@ -62,15 +62,15 @@ query_cc_antibodies("CD11c")
 # 8 Antibody Conjugates
 
 query_sc_antibodies("CD11bfakename")
-NULL
+# NULL
 head(query_sc_antibodies("CD11C"))
 # Product_Name  Cat_num Citations Rating     Epitope
-# 1   Integrin alpha X Antibody (B-6) sc-46676         3    (8) Integrin αX
-# 2   Integrin alpha X Antibody (3.9)  sc-1185         2    (2) Integrin αX
-# 3 Integrin alpha X Antibody (B-Iy6) sc-19989         1    (1) Integrin αX
-# 4  Integrin alpha X Antibody (N418) sc-23951         3    (2) Integrin αX
-# 6 Integrin alpha X Antibody (2Q865) sc-71455       NEW    (1) Integrin αX
-# 7 Integrin alpha X Antibody (3H986) sc-71456       NEW    (1) Integrin αX
+# 1   Integrin alpha X Antibody (B-6) sc-46676         3    (8) Integrin aX
+# 2   Integrin alpha X Antibody (3.9)  sc-1185         2    (2) Integrin aX
+# 3 Integrin alpha X Antibody (B-Iy6) sc-19989         1    (1) Integrin aX
+# 4  Integrin alpha X Antibody (N418) sc-23951         3    (2) Integrin aX
+# 6 Integrin alpha X Antibody (2Q865) sc-71455       NEW    (1) Integrin aX
+# 7 Integrin alpha X Antibody (3H986) sc-71456       NEW    (1) Integrin aX
 # Species                       Method
 # 1   human WB, IP, IF, IHC(P) and ELISA
 # 2   human               IP, IF and FCM
