### 2020-02-04 jspaezp/sctree#3 @jspaezp
I have pushed several commits that address these issues.
Currently all checks and tests pass on my desktop and docker environments.
I will leave this open until it successfully passes the tests in Travis
