### 2020-02-04 f8b0fae
updated webpage, travis and removed explicit import in example
