+++ R/utils.R
@@ -350,7 +350,6 @@ autoplot <- function(object, ...) {
 #' @export
 #'
 #' @examples
-#' require(ggplot2)
 #' autoplot(table(1:10, 1:10), show_number = TRUE)
 #' @importFrom ggplot2 geom_tile geom_point aes_string labs autoplot
 autoplot.table <- function(object,
