+++ .travis.yml
@@ -9,5 +9,17 @@ r:
   - bioc-devel
   - bioc-release
 
+before_cache: Rscript -e 'remotes::install_cran("pkgdown")'
+
 after_success:
   - Rscript -e 'covr::codecov()'
+  - Rscript -e 'pkgdown::build_site()'
+
+deploy:
+  provider: pages
+  skip-cleanup: true
+  github-token: $GITHUB_PAT
+  keep-history: true
+  local-dir: docs
+  on:
+    branch: master
