### 2020-02-04 jspaezp/sctree#3 @jspaezp
@mschubert @natallah
Tests have passed on the `bioc-release` and `bioc-devel` branches! There were indeed a couple of things that needed fixing due to a change in the request structure for santa cruz antibodies and an issue where special characters made latex fail when generating the manual.

https://travis-ci.org/jspaezp/sctree/jobs/645977617

please let me know if any additional changes are required.
