### 2020-02-04 1d87267
removed unused files and rebuilt documentation
