+++ templates/benchmarking_ml.Rmd
@@ -1,448 +1,0 @@
----
-title: "ML- algs testing"
-author: "J. Sebastian Paez"
-date: "`r Sys.Date()`"
-output: 
-  html_document: 
-  rmdformats::material:
-    highlight: kate
-    keep_md: yes
-params:
-    SEURAT_RDS: 'seurat_mix5050.RDS'
-    MARKERS_RDS: 'markers_mix5050.RDS'
-    GLOBAL_REPEATS: 1
-    NUM_CROSS_VAL: 2
-    DATA_FRACTION: 0.01
-    FILTER_MEMBRANE: TRUE
-    NUMTHREADS: 2
-    DATASET_ID: 'somerandomname'
-    CLUSTER: 'X1'
-
----
-
-
-```{r knitr_init, echo=FALSE, cache=FALSE}
-library(knitr)
-library(rmdformats)
-
-## Global options
-options(max.print = "75")
-opts_chunk$set(echo = TRUE,
-	             cache = FALSE,
-               prompt = FALSE,
-               tidy = TRUE,
-               comment = NA,
-               message = TRUE,
-               warning = TRUE)
-opts_knit$set(width = 75)
-```
-
-# Setup
-
-```{r}
-warning(str(params))
-```
-
-
-```{r}
-library(doSNOW)
-cl <- makeCluster(params$NUMTHREADS, type = "SOCK")
-registerDoSNOW(cl)
-
-```
-
-```{r message=FALSE, warning=FALSE, paged.print=FALSE}
-library(Seurat)
-library(ggplot2)
-library(plyr)
-library(dplyr)
-library(ranger)
-library(party)
-library(caret)
-library(org.Hs.eg.db)
-```
-
-```{r message=FALSE, warning=FALSE, paged.print=FALSE}
-print(params)
-dataset <- readRDS(params$SEURAT_RDS)
-markers <- readRDS(params$MARKERS_RDS)
-
-set.seed(42)
-```
-
-# Helper Functions
-
-```{r}
-#' Gets a data frame from a seurat object
-#' 
-#' Provided a seurat object, returs a data frame of the count values, being 
-#' the columns each 'gene' and the rows each UMI/cell.
-#'
-#' It returns only the genes annotated as variable and the identity column.
-#'
-#' @param seurat A seurat object
-#'
-#' @return
-#' @export
-#'
-#' @examples
-#' > as.data.frame.Seurat(dataset.minus_small)[1:3,1:3]
-#'                    LTB EAF2 CD19
-#' ATGCCAGAACGACT 6.062788    0    0
-#' CATGGCCTGTGCAT 6.714813    0    0
-#' GAACCTGATGAACC 7.143118    0    0
-#' @importFrom Seurat FetchData GetIdent
-as.data.frame.Seurat <- function(seurat) {
-    tmp <- Seurat::FetchData(seurat, vars.all = seurat@var.genes)
-    tmp <-  as.data.frame(tmp, stringsAsFactors = FALSE)
-
-    tmp$ident <- Seurat::GetIdent(
-      seurat, uniq = FALSE, cells.use = rownames(tmp))
-    return(tmp)
-}
-
-
-dataset.df <- as.data.frame.Seurat(dataset)
-
-# Some minor renaming is required for most methods, since some genes have - in 
-# their names, and we require R-valid names.
-
-original_names <- colnames(dataset.df)
-newnames <- make.names(original_names)
-colnames(dataset.df) <- newnames
-dim(dataset.df)
-```
-
-
-```{r}
-
-# The markers df object is supposed to be comming form Seurat's FindAllMarkers
-fit_models <- function(forest_topgenes, 
-                       wilcox_topgenes,
-                       wilcox_topgenes_local,
[…]
