### 2020-02-05 jspaezp/sctree#4 @jenzopr
[joss review] Installation functionality
Hi @jspaezp!

This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2061).

Can you please update the installation instructions in the README(.md|.Rmd)? The URL to the Git repository given there is not valid anymore. Also, consider switching to `remotes::install_github` instead of a `git checkout` and `devtools::install`.

Using `remotes::install_github("jspaezp/sctree")` works fine and without errors for me.
