### 2020-02-05 jspaezp/sctree#5 @jenzopr
[joss review] Documentation: Example usage
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2061).

## Summary

The example usage sections (entitled with *Usage* and *Seurat Interface* in the README.md file) contain detailed code examples for getting started with the package and are fine-grained enough to easily follow along. I was able carve out the following basic analysis workflow:

1. Find marker genes
2. (optional) Visualize outcome of expected flow cytometry experiment
3. Get and validate a suggested gating strategy
4. Visualize expected flow cytometry experiment and gating
5. Find antibodies

However, the presented order of functions complicate understanding of the intended analysis workflow and needs improvement (see major and minor issues below).

## Major issues

* Three functions are presented to find marker genes, namely `ranger_importances.Seurat`, `findAllMarkers_ranger` and `FindAllMarkers` (together with its complementary function `FindMarkers`). What are the differences between those and when should I use which? Should or can those functions be used complimentary?
* Having obtained a suggested gating strategy with `fit_ctree`, how can quality of the fitted tree be assessed? Which measurement(s) can be used to evaluate the *goodness* of the tree?

## Minor issues encountered during testing

* `RangerDE` is not a valid option for `test.use` in `sctree::FindMarkers()`. The error message is `Unknown test: RangerDE`.
* `sctree::FindMarkers()` fails with `Error in PackageCheck("MAST", error = FALSE): could not find function "PackageCheck"`, when using `test.use = "MAST"` and MAST is not installed. Consider using `requireNamespace("MAST", quietly = T)` to check if MAST is installed (Ref: http://r-pkgs.had.co.nz/namespace.html). If not, print an informative message to the user that MAST is required for this function.
* The output of `sctree::plot_flowstyle` easily becomes non-readable, when there are more than 8 clusters / colors. IMHO the possibility to *highlight* cells from a cluster of interest would enhance the usage with many clusters.
