### 2020-02-07 jspaezp/sctree#5 @jspaezp
Hello @jenzopr

## Looking at the major issues:

    Three functions are presented to find marker genes, namely ranger_importances.Seurat,
    findAllMarkers_ranger and FindAllMarkers (together with its complementary function FindMarkers).
    What are the differences between those and when should I use which? Should or can those
    functions be used complimentary?

I could specify in the documentation that `FindAllMarkers` and `FindMarkers` are the high-level functions that most users would be interested in and `ranger_importances.Seurat` and `findAllMarkers_ranger` are internals that are usually called by these high-level functions. I feel like keeping these functions separate is essential to maintain consistency with the equivalent functions in `Seurat` but the addition of details in the documentation and vignettes would be useful.

    Having obtained a suggested gating strategy with fit_ctree, how can quality of the fitted tree be
    assessed? Which measurement(s) can be used to evaluate the goodness of the tree?

Since the object returned has a `predict` method, It could be used to generate a prediction based on a testing/training split on the data. A vignette explaining how to carry that out would be really useful for the end-user.

## Looking at the minor issues:

1. I noticed that calling `sctree::FindMarkers` fails only when called before loading the pacakge, which might have to do with how calls its internals. I feel like the most straightforward way to fix it would be to test if `WhichCells` is in the search path and throw a more informative error saying that it needs to be loaded.

```r
sctree::FindMarkers(
  object = Seurat::pbmc_small,
  ident.1 = 0,
  ident.2 = 1,
  test.use = "RangerDE",
  verbose = FALSE)
# Error in WhichCells(object = object, idents = ident.1) :
#   could not find function "WhichCells"
# Error in inherits(ok, "try-error") : object 'ok' not found

require(sctree)

sctree::FindMarkers(
  object = Seurat::pbmc_small,
  ident.1 = 0,
  ident.2 = 1,
  test.use = "RangerDE",
  verbose = FALSE)
#                  importance      p_val          gene  avg_logFC pct.1
# GNLY           0.0530090754 0.00000000          GNLY  5.1338001 0.444
# CD7            0.0961302279 0.00000000           CD7  4.5355550 0.528
# PRF1           0.0436675892 0.00000000          PRF1  4.0809632 0.417
#................
#................
#................
# IL17RA         0.20         1
# SNHG7          0.12         1
# RBP7           0.28         1

```

2. Regarding the function throwing an error when MAST is not installed, that is due to how the function handles the missing package in Seurat, nonetheless, it would be easy enough to just add specific error handling for it. Nonetheless, the error propagates well enough to give an informative error message (will add the exception).

3. Regarding the output of `sctree::plot_flowstyle`, I agree with it being a good idea, will definitely implement that. I guess that handling it in a similar way that `plot_gates` does it would be good.

# TODO

- [x]  documentation specifying the intended use of `FindAllMarkers`, `FindMarkers`, `ranger_importances.Seurat`, `findAllMarkers_ranger` as high or low level functions.
- [x] testing/training split vignette
- [x] test if `WhichCells` is in the search path when calling `sctree::FindMarkers`
- [x] add the exception when MAST is not installed
- [x] add cluster highlighting to `plot_flowstyle`
