### 2020-02-07 jspaezp/sctree#3 @jspaezp
Will mark as closed, since all the builds complete successfully in travis.
