### 2020-02-07 af0b033
added "remotes" as an alternative installation option.
