+++ .travis.yml
@@ -1,7 +1,6 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
-language: R
-sudo: false
+language: r
 cache: packages
 warnings_are_errors: false
 
@@ -17,9 +16,9 @@ after_success:
 
 deploy:
   provider: pages
-  skip-cleanup: true
-  github-token: $GITHUB_PAT
-  keep-history: true
-  local-dir: docs
+  cleanup: false
+  token: $GITHUB_PAT
+  keep_history: true
+  local_dir: docs
   on:
     branch: master
