### 2020-02-07 39b00cc
fixed rcac link in the installation instructions
