### 2020-02-07 jspaezp/sctree#4 @jspaezp
Hello @jenzopr

Thank you for noticing this!

addressed it in 39b00cc09be9e800f22dbcbb5fb5710149ca0704
(also created a hotfix branch for the fixes involved in the review)

Kindest wishes,
Sebastian
