### 2020-02-12 jspaezp/sctree#4 @mschubert
I'm a bit confused about your installation instructions. Why do you want your users to clone manually and then install the package using `devtools`?

Wouldn't it be easier to do either a

```r
devtools::install_github("jspaezp/sctree")
```

or

```r
remotes::install_github("jspaezp/sctree")
```

without the manual cloning & installing steps?
