### 2020-02-12 jspaezp/sctree#2 @mschubert
This may seem a bit nitpicking (and is not required for JOSS, so feel free to ignore), but note that the file name `CONTIRUBUTING.md` (upper/lower case is I think irrelevant) will be recognized by Github (and other machine parsers) as contribution guidelines. `Contribute.md` will not.

If you click on Insights>Community, you will see what I mean.
