### 2020-02-12 jspaezp/sctree#6 @mschubert
[joss review] dependencies
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2061).

For ticking off the point

> **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

I still see some unstated dependencies (not listed in `DESCRIPTION`):

* [DESeq2](https://github.com/jspaezp/sctree/blob/e69c96b14e1adc378d1f78865882c31fdb8741fe/R/seurat_utils.R#L968)
* [MAST](https://github.com/jspaezp/sctree/blob/e69c96b14e1adc378d1f78865882c31fdb8741fe/R/seurat_utils.R#L1373)
* [RColorBrewer](https://github.com/jspaezp/sctree/blob/6cf91ea525e6a99fcbf80ab5f75b27f2f509b45e/R/print_tree_rules.R#L477)
* [SummarizedExperiment](https://github.com/jspaezp/sctree/blob/e69c96b14e1adc378d1f78865882c31fdb8741fe/R/seurat_utils.R#L1404-L1406)

These are also listed as warnings in your Travis package check.

(Note that these can be listed in `Suggests:` if they are optional)
