### 2020-02-12 jspaezp/sctree#5 @mschubert
> ranger_importances.Seurat and findAllMarkers_ranger are internals that are usually called by these high-level functions

If they are internals, are you sure you want to [`@export`](https://github.com/jspaezp/sctree/blob/e69c96b14e1adc378d1f78865882c31fdb8741fe/R/ranger_utils.R#L193) them?

You can also mark them with `@keywords internal` so they don't show up in the API documentation index and are less prone to confuse users.
