### 2020-02-14 jspaezp/sctree#7 @jenzopr
[joss review] Software paper
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2061).

## Software paper review

In their paper, Paez et. al. identify unfulfilled needs during the translation of findings from computational scRNA-seq analysis to downstream wet-lab methodologies. Especially for marker gene detection, a critical step in characterization of cellular (sub)populations,  there seems to be no applicable method that transfers knowledge from *in silico* to the bench. They present *scTree*, an R package for marker gene detection that employs random forests for variable selection and a classification tree that resembles FACS gating strategies, thereby enhancing interpretability and application of detected marker genes in downstream wet-lab experiments. Overall the manuscript is well written and does not require major editing for structure or language. They benchmark the quality of their method quite elegantly using recall statistics on test data that has been left out during training.

#### Major issues

- **Target audience**: From the introduction, it becomes somewhat clear that *scTree* is a tool/package, but the anticipated target audience (e.g. (computational) biologists with knowledge of the R programming language and Seurat) does not become apparent. Please improve wording of the last paragraph of the introduction to specify the target audience.
- **State of the field**: Authors describe a range of tools in the field that *fail to directly connect findings [...] to be used downstream*. In my view, this comparison is problematic, since many of the tools mentioned to not claim to achieve that. For example, *short sequence alignment*, *barcode calling*, *preprocessing*, *clustering* and *differential gene expression* are all **needed** in the first place in order to be able to transfer findings to the bench. Please improve wording of this paragraph and add a more in-depth analysis of tools that claim to find biologically relevant and interpretable marker genes.
- Figure legends could be enhanced to increase understanding of the figures. E.g. it is not clear to me, where predicted identities in Figure 2 come from, what message they convey and how those predictions integrate into the logic of the benchmark. Further, to measure similarity of a prediction and the real identity, the adjusted rand index could be used.

#### Minor issues

- Page 1, paragraph 3: A short explanation of **what marker genes are** is missing
- Page 2, paragraph 3: random forests **has** been previously
