### 2020-02-17 90d0d51
added explicit import of WhichCells and changed how the marker finding functon finds for arbitrary function passing
