### 2020-02-17 0cbc8f4
Fixed namespace import of differential expression functions.
