### 2020-02-17 4ad31e0
added option to highlight clusters in plowstyle_plot
