### 2020-02-17 jspaezp/sctree#5 @jspaezp
Update:
1. Cluster highlighting implemented in 4ad31e0474f46ce34b2525fa2e1b76758c652a51
2. Fixed issue where MAST (and deseq2) would not be flagged correctly as missing installation 0cbc8f4dec192947ec04da968eada1c350de41c0
3. Added better handling of WhichCells, having it be correctly imported from Seurat 90d0d517cdb039657cce87c766a726abc7cc827a

@natallah
