### 2020-02-17 8ee7dd9
added optional packages to suggests section
