### 2020-02-17 jspaezp/sctree#2 @jspaezp
hello @mschubert, I have added this change in d2d85d41d014b407bdb632dd2d4f0d9543b9738e
