### 2020-02-17 jspaezp/sctree#4 @jspaezp
Hello @mschubert !

I add that option in case the user wants to modify the repository and install the modified version. Indeed using remotes is the easiest way to do it.

Do you believe it would be required for me to remove it? Since it is a fairly restricted use-case
