### 2020-02-17 jspaezp/sctree#4 @mschubert
Your current instructions are valid, so keeping your manual cloning and installation to will not block your paper's acceptance (so this is your decision).

However, I don't understand your reasoning. I'd think the installation instructions are there primarily to help novice users; more experienced package developers (who are the ones that may want to change your code) will already know that they can clone manually.
