### 2020-02-17 jspaezp/sctree#3 @mschubert
@jspaezp your build seems to be failing again: https://travis-ci.org/jspaezp/sctree/builds/651358734
