### 2020-02-17 jspaezp/sctree#8 @mschubert
[joss review] documentation
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2061).

I see that you have compiled your documentation into a website: https://jspaezp.github.io/sctree/

Please also link this as Github URL and/or from the README, otherwise your users will not find it.
