### 2020-02-17 jspaezp/sctree#8 @mschubert
Minor webpage: the "home" and "getting started" links link to the same information; provide this only once to not confuse users?
