### 2020-02-17 jspaezp/sctree#7 @mschubert
I agree with the issues @jenzopr raised.

In addition, I would like to see improvements for the following parts:

* The software paper is quite verbose and takes a while to come to a point (more akin to a "tranditional" journal introduction compared to other JOSS papers). As far as I see it, the main novelty is that `scTree` suggests a (small) set of genes that separate populations. Why not state this right in the beginning?
* I do not see how references to "other people did scRNA-seq alignment [etc.]" are relevant for this software paper. Doesn't it make more sense to focus on what this software does? (i.e., find sortable marker genes between populations)
* Based on the previous points, the "summary" part of the introduction could be simplified to something like this?:

  > Single-cell RNA sequencing (scRNA-seq) is a now commonly used technique to measure the transcriptome of many cells. Clusters of these transcriptomes identify cell populations (ref). There are multiple methods available to identify "marker" genes that separate these populations (refs). However, there are usually too many genes in these lists to directly suggest an experimental follow-up strategy for selecting them from a bulk population (e.g. via FACS (ref)). Here we present `scTree`, a tool that aims to provide a minimal set of genes to separate populations in scRNA-seq in a follow-up experiment.
* Implementation, citations for `ranger` package: only one is the package citation, are the others relevant for this software paper? If I as a reader am particularly interested in how `ranger` implements their RF, I'd look at the citations in that paper
* Implementation, benchmark: I'm confused about what you are looking at in your benchmark. Doesn't `scTree` use RF? Why is it listed with 6 different ML methods? (are those available but RF is default?) Where does the t-test and the Wilcox come in? This is not clear from the software paper.
* Implementation, benchmark: You make the point that previous marker gene methods are not suitable because there are too many genes that can't be used in a sorting experiment. Hence, wouldn't a better benchmark a different marker gene method (& then using e.g. top 5 genes) vs. `scTree` to underline this point?
* Minor: Reference "Satija-Lab" should probably contain a URL, ideally via https://archive.org/web/ to prevent link rot
* Minor: typo "clasification"
* Minor: typo "ha been previously"
