### 2020-02-17 jspaezp/sctree#8 @mschubert
https://jspaezp.github.io/sctree/reference/index.html

API documentation could also be improved: The title of many functions is equal to their name. This does not help a user to know what these functions are doing. Maybe provide more descriptive titles? (and do not list internal functions there)
