### 2020-02-17 jspaezp/sctree#8 @mschubert
Also minor README/docs: You are `require(...)` to load packages, but do not check for a return value.

This may lead to confusing errors for users if the package is not installed: `require` will only warn, and the script will produce an error elsewhere (that may not sound like the source is a missing package). If you use `library(...)` instead, a script will terminate with an error if the package is not installed.
