### 2020-02-17 jspaezp/sctree#8 @jspaezp
Hello @mschubert, I can definitely agree with all of these, I am compiling this into a todo list.
Will keep you posted as I go though it.

TODO:

- [x] link this as Github URL and/or from the README, otherwise your users will not find it.
- [x] the "home" and "getting started" links link to the same information. Fix it.
- [x] provide more descriptive titles in the documentation
- [x] change `require` to `library` in documentation.

@natallah
