### 2020-02-24 b402fbe
api clarifications, version bump and restructuring of the reference section in the companion website. Also, modification of internals as S3 methods and flagging several functions as internal
