+++ README.md
@@ -19,9 +19,12 @@ Features suggesting pseudo-gating strategies to purify found populations
 via flow-cytometry, antibody querying and cross validations between
 datasets.
 
-Number of lines in roxygen comments: 1024
+Please check our companion documentation website hosted at
+[jspaezp.github.io/sctree](https://jspaezp.github.io/sctree/).
 
-Number of lines in R code: 1946
+Number of lines in roxygen comments: 1039
+
+Number of lines in R code: 1968
 
 # Installation
 
@@ -60,8 +63,7 @@ Original data can be found here:
 <!-- end list -->
 
 ``` r
-require(sctree)
-#> Loading required package: sctree
+library(sctree)
 #> Loading required package: Seurat
 #> Registered S3 method overwritten by 'GGally':
 #>   method from   
@@ -71,7 +73,7 @@ require(sctree)
 #> The following objects are masked from 'package:Seurat':
 #> 
 #>     FindAllMarkers, FindConservedMarkers, FindMarkers
-require(Seurat)
+library(Seurat)
 
 set.seed(6)
 
@@ -102,7 +104,7 @@ when addressing this issue and for more details.
 
 ``` r
 library(sctree)
-rang_importances <- ranger_importances.Seurat(
+rang_importances <- ranger_importances(
     small_5050_mix,
     cluster = "ALL",
     warn.imp.method = FALSE)
@@ -125,21 +127,24 @@ head(rang_importances)
 # Seurat Interface
 
 As an analogous function to Seurat’s `FindAllMarkers`, we offer
-`FindAllMarkers_ranger.Seurat`
+`ranger_importances` or the `RangerDE` option for `FindAllMarkers`
 
 ``` r
-markers <- FindAllMarkers_ranger.Seurat(
+markers <- FindAllMarkers(
   small_5050_mix,
-  warn.imp.method = FALSE)
+  warn.imp.method = FALSE, 
+  test.use = "RangerDE")
+#> Calculating cluster 0
+#> Calculating cluster 1
 
 head(markers)
-#>         importance pvalue    gene cluster
-#> ASNS      5.685892      0    ASNS       0
-#> TMSB4X    5.403575      0  TMSB4X       0
-#> ARHGDIB   5.114462      0 ARHGDIB       0
-#> ADA       3.853724      0     ADA       0
-#> CD3D      3.558077      0    CD3D       0
-#> MZB1      3.310409      0    MZB1       0
+#>         importance p_val    gene avg_logFC pct.1 pct.2 p_val_adj cluster
+#> MZB1     0.9641974     0    MZB1  2.004299 0.828 0.227         0       0
+#> CD3G     1.2004702     0    CD3G  1.849551 0.744 0.107         0       0
+#> CD3D     3.2936540     0    CD3D  1.842246 0.872 0.373         0       0
+#> ARHGDIB  2.3889365     0 ARHGDIB  1.787994 0.856 0.320         0       0
+#> AIF1     0.4203789     0    AIF1  1.759311 0.756 0.173         0       0
+#> FYB      0.8291528     0     FYB  1.750418 0.739 0.107         0       0
 
 plot.markers <- do.call(rbind, lapply(split(markers, markers$cluster), head, 3))
 
@@ -389,7 +394,7 @@ Here is a simple example for a gene widely know to have an antibody
 available
 
 ``` r
-require(sctree)
+library(sctree)
[…]
