### 2020-02-24 jspaezp/sctree#8 @jspaezp
Commit b402fbe9eaaa17c352053f65c953b173c4f0e5f5 adds the reference to the website, and changes some of the titles in the reference section of the documentation. It also changes in all vignettes the `require` for the `library` function.

@natallah and @mschubert
