### 2020-02-24 jspaezp/sctree#5 @jspaezp
Hello @jenzopr and @mschubert

Commit b402fbe9eaaa17c352053f65c953b173c4f0e5f5 further includes changes on this regard with the clarification in the documentation on what functions are intended for end-user and which are internals, indeed it was a good idea to have `@keywords internal` added.

@natallah
