### 2020-02-24 6a5fda3
Moved the manual isntallation instructions to the contributing.md file from the readme
