+++ vignettes/rmdchunks/_installation.Rmd
@@ -8,13 +8,3 @@ installed by using the following function in R
 ```r
 remotes::install_github("jspaezp/sctree")
 ```
-
-
-Alternatively, one can clone the git repository to test changes.
-
-```
-git clone https://github.com/jspaezp/sctree sctree
-cd sctree
-
-R -e "devtools::install('.')"
-```
