### 2020-02-24 jspaezp/sctree#4 @jspaezp
Hello @mschubert,

after thinking about it i decided to move the manual install instructions to the contributions file from the readme. Changes are done in 6a5fda3075217a691809019afdc612b18c44306c

Kindest wishes,
Sebastian

@natallah
