### 2020-02-24 jspaezp/sctree#3 @jspaezp
Looking at the log, it seems like the part failing is not in my tests on *bioc-devel* ...

```
* installing *source* package ‘GenomeInfoDb’ ...

** using staged installation

** R

** inst

** byte-compile and prepare package for lazy loading

Error: package or namespace load failed for ‘S4Vectors’:

 .onLoad failed in loadNamespace() for 'S4Vectors', details:

  call: validObject(.Object)

  error: invalid class “LLint” object: superclass "integer_OR_LLint" not defined in the environment of the object's class

Error: package ‘S4Vectors’ could not be loaded

Execution halted

ERROR: lazy loading failed for package ‘GenomeInfoDb’

* removing ‘/home/travis/R/Library/GenomeInfoDb’

Error in i.p(...) :

  (converted from warning) installation of package ‘GenomeInfoDb’ had non-zero exit status

Calls: <Anonymous> ... with_rprofile_user -> with_envvar -> force -> force -> i.p

Execution halted
```
