+++ NAMESPACE
@@ -24,6 +24,7 @@ export(get_aliases)
 export(get_cluster_mapping)
 export(get_concensus_rules)
 export(get_genesymbols)
+export(ggheatmap_base)
 export(is.confusion.matrix)
 export(is_gene_membrane)
 export(plot_flowstyle)
