### 2020-02-24 jspaezp/sctree#3 @jspaezp
on the other hand the "bioc-stable" release works fine

https://travis-ci.org/jspaezp/sctree/jobs/654457775
