### 2020-03-09 ffa3008
moved vignettes to prevent automatic insertion in pkgdown website and re-built documentation
