+++ news.md
@@ -1,0 +1,13 @@
+
+# sctree 0.0.8
+
+## Major changes
+
+- Working on the revision to get the paper as part of JOSS
+
+## Minor changes
+
+- Removed redundancy of the readme vignette and the getting started for
+  the pkgdown website.
+- Added vignette explaining how to train and check the performance of
+  a classification tree
