+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: sctree
 Title: Tree based analysis of single rna-seq clusters
-Version: 0.0.7.0000
+Version: 0.0.8.0000
 Authors@R: 
     person(given = "J. Sebastian",
            family = "Paez",
