### 2020-03-09 96a1c9a
Modify paper.md
Specified target audience and simplified the last paragraph of paper.md.
