### 2020-03-09 9414876
Changed intro
Simplified introduction paragraph.  Specifies target audience, removes irrelevant information, and cites commonly used packages to identify marker genes.  Need to add a FACS reference still.
