### 2020-03-09 df18f53
Update bibliography.bib
Updated bibliography to reflect changes to the Introduction
