+++ vignettes/paper/bibliography.bib
@@ -55,82 +55,7 @@ year = {2017}
   pages = {3905-3909},
   url = {http://jmlr.org/papers/v16/hothorn15a.html},
 }
-@article{Srivastava2018,
-abstract = {We introduce alevin, an efficient pipeline for gene quantification from dscRNA-seq (droplet-based single-cell RNA-seq) data. Alevin is an end-to-end quantification pipeline that starts from sample-demultiplexed FASTQ files and generates gene-level counts for two popular droplet-based sequencing protocols (drop-seq [1], and 10x-chromium [2]). Importantly, alevin handles all processing internally, avoiding reliance on external pipeline programs, and the need to write large intermediate files to disk. Alevin adopts efficient algorithms for cellular-barcode whitelist generation, cellular-barcode correction, lightweight per-cell UMI deduplication and quantification. This integrated solution allows alevin to process data much faster (typically {\~{}} 10 times faster) than other approaches, while also working within a reasonable memory budget. This enables full, end-to-end analysis for single-cell human experiment consisting of {\~{}} 4500 cells with 335 Million reads with 13G of RAM and 8 threads (of an Intel Xeon E5-2699 v4 CPU) in 27 minutes.},
-author = {Srivastava, Avi and Smith, Tom Sean and Sudbery, Ian and Patro, Rob},
-doi = {10.1101/335000},
-journal = {bioRxiv},
-keywords = {cellular barcode,dscrna-seq,quantification,umi},
-mendeley-groups = {Sequencing},
-pages = {335000},
-title = {{Alevin: An integrated method for dscRNA-seq quantification}},
-url = {https://www.biorxiv.org/content/early/2018/06/01/335000},
-year = {2018}
-}
-@article{Patro2017,
-abstract = {We introduce Salmon, a method for quantifying transcript abundance from RNA-seq reads that is accurate and fast. Salmon is the first transcriptome-wide quantifier to correct for fragment GC content bias, which we demonstrate substantially improves the accuracy of abundance estimates and the reliability of subsequent differential expression analysis. Salmon combines a new dual-phase parallel inference algorithm and feature-rich bias models with an ultra-fast read mapping procedure. },
-author = {Patro, Rob and Duggal, Geet and Love, Michael I and Irizarry, Rafael A and Kingsford, Carl},
-doi = {10.1038/nmeth.4197},
-journal = {Nature Methods},
-mendeley-groups = {Sequencing},
-number = {4},
-pages = {417--419},
-title = {{Salmon: fast and bias-aware quantification of transcript expression using dual-phase inference}},
-volume = {14},
-year = {2017}
-}
-@article{Li2018,
-abstract = {The emerging single-cell RNA sequencing (scRNA-seq) technologies enable the investigation of transcriptomic landscapes at the single-cell resolution. ScRNA-seq data analysis is complicated by excess zero counts, the so-called dropouts due to low amounts of mRNA sequenced within individual cells. We introduce scImpute, a statistical method to accurately and robustly impute the dropouts in scRNA-seq data. scImpute automatically identifies likely dropouts, and only perform imputation on these values without introducing new biases to the rest data. scImpute also detects outlier cells and excludes them from imputation. Evaluation based on both simulated and real human and mouse scRNA-seq data suggests that scImpute is an effective tool to recover transcriptome dynamics masked by dropouts. scImpute is shown to identify likely dropouts, enhance the clustering of cell subpopulations, improve the accuracy of differential expression analysis, and aid the study of gene expression dynamics.},
-author = {Li, Wei Vivian and Li, Jingyi Jessica},
-doi = {10.1038/s41467-018-03405-7},
[…]
