+++ vignettes/paper/bibliography.bib
@@ -55,7 +55,19 @@ year = {2017}
   pages = {3905-3909},
   url = {http://jmlr.org/papers/v16/hothorn15a.html},
 }
-
+@article{RN1,
+   author = {Tung, J. W. and Heydari, K. and Tirouvanziam, R. and Sahaf, B. and Parks, D. R. and Herzenberg, L. A. and Herzenberg, L. A.},
+   title = {Modern flow cytometry: a practical approach},
+   journal = {Clin Lab Med},
+   volume = {27},
+   number = {3},
+   pages = {453-68, v},
+   ISSN = {0272-2712 (Print)
+0272-2712},
+   DOI = {10.1016/j.cll.2007.05.001},
+   year = {2007},
+   type = {Journal Article}
+}
 @article{Love2014,
 abstract = {In comparative high-throughput sequencing assays, a fundamental task is the analysis of count data, such as read counts per gene in RNA-seq, for evidence of systematic changes across experimental conditions. Small replicate numbers, discreteness, large dynamic range and the presence of outliers require a suitable statistical approach. We present DESeq2, a method for differential analysis of count data, using shrinkage estimation for dispersions and fold changes to improve stability and interpretability of estimates. This enables a more quantitative analysis focused on the strength rather than the mere presence of differential expression. The DESeq2 package is available at http://www.bioconductor.org/packages/release/bioc/html/DESeq2.html webcite.},
 author = {Love, Michael I and Huber, Wolfgang and Anders, Simon},
