### 2020-03-09 a5e79d0
Minor update of paper.md
Added FACS review reference to the introduction.   Addressed minor issues from Issue #7
