### 2020-03-09 41e3866
Updated Figure legend 1 in paper.md
Enhanced Figure legend 1.  Also added rationale for testing with various classifiers.
