### 2020-03-09 417c44b
Update Figure legend 2 in paper.md
Added an improved legend for Figure 2 and an explanation of what "Real Identify".
