### 2020-03-09 20f9bf1
Updated paper.md
Added reasoning for why scTree was compared with t-test and Wilcoxon tests
