### 2020-03-09 jspaezp/sctree#7 @natallah
Thank you @jenzopr and @mschubert for your comments!  I have addressed the major and minor issues .
