### 2020-03-09 dc788fd
added descriptive titles to documentation and updated website
