### 2020-03-09 jspaezp/sctree#5 @jspaezp
@jenzopr @natallah @mschubert

Aadded train/test vignette in commit ffa30082ed60cc7dc0d9bdd26197ebaedbaa4bfe

This completes the checklist for this issue, let me know if you believe we can close this issue.
Kindest wishes,
Sebastian
