### 2020-03-09 jspaezp/sctree#8 @jspaezp
Home and getting started section in the pkgdown website have been resolved in ffa30082ed60cc7dc0d9bdd26197ebaedbaa4bfe

Descriptive titles to the documentation in the reference section is added on commit dc788fd33839397c5528148b7a6b8c4da276a62b

This completes the TODO list for this issue, so I will go ahead and close it.
@mschubert @natallah
