### 2020-03-10 e741368
added documentation and fixed gitignore
