### 2020-03-10 jspaezp/sctree#5 @jenzopr
Very nice! I really like the `Evaluating Tree Quality` vignette!
