### 2020-03-10 jspaezp/sctree#7 @jenzopr
From my point of view, the paper gained substantially from your edits! Nice work :+1:
