### 2020-03-13 jspaezp/sctree#5 @jenzopr
A last small request from my side: Can you update the introduction of the `README.md` file and include a bit more from the papers Summary to fulfill the **statement of need** from the JOSS checklist: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
You basically have this nicely in the paper, but not yet in the Readme and webpage.
Thanks!
