### 2020-03-13 f82dbfd
README.md
Added a statement of need to the README.md
