### 2020-03-13 jspaezp/sctree#5 @natallah
@jspaezp @jenzopr

I updated the README.md accordingly!  Great suggestion - thanks!
