### 2020-03-14 #2061 @natallah
@jenzopr   @jspaezp
The proof overall looks really good!  One thing we  need to figure out is why Figure 3 in the paper.md is showing up labeled as Figure 1 in the pdf.  We definitely do not want to have two figure 1s!
