### 2020-03-14 56413c2
Update Figure 3 legend in paper.md
Added legend to paper.md
