### 2020-03-16 3aa2e98
added new dfigures to the paper and minor modifications on the text
