+++ vignettes/paper/paper.md
@@ -66,13 +66,13 @@ of the prediction. We were concerned that the forest-based markers would artific
 favor scTree, therefore we utilized several classifiers for the markers derived from
 either scTree, t-tests or wilcoxon-tests.
 As shown in **Figures 1 and 2**, bias was not observed, and regardless of the final
-classification model, the features selected by using scTree provide a comparable accuracy, precision and recall to those acquired using traditional differential expression methods.
+classification model, the features selected by using scTree provide a comparable accuracy, precision and recall to those acquired using traditional differential expression methods. It is to note that many of the wrongly assigned labels happen between cell populations that are hard to define in-vivo and are not resolved clusters in the UMAP dimensional reduction, such as macrophage subtypes and between NK an Tc cells.
 
 ![Depiction of the classification performance achieved in the Jurkat:293 50:50 dataset](./paper_figures/benchmarks_5050_boxplot.png "Figure 1")
 **Figure 1. Depiction of the classification performance achieved in the Jurkat:293 50:50 dataset.**  A number of machine learning algorithms were tested to ensure that scTree performed as well as traditional marker identification approaches, regardless of the classifier used.
 
-![Depiction of the predicted identities in the PBMC 3k dataset dataset.](./paper_figures/pbmc_3k_prediction_UMAP.png "Figure 2")
-**Figure 2. Depiction of the predicted identities in the PBMC 3k dataset dataset.** Real identities are the identities as identified through unsupervised clustering performed using Seurat [@Butler2018].  The scTree package was then used to then classify cells based on the top 5 markers for each cluster chosen by scTree and accurately recapitulates the original classification determined by Seurat. 
+![Depiction of the predicted identities in the PBMC 3k dataset dataset.](./paper_figures/pbmc_prediction_sctree.png "Figure 2")
+**Figure 2. Depiction of the predicted identities in the PBMC 3k dataset dataset.** Real identities are the identities as identified through unsupervised clustering performed using Seurat and annotated based on prior knowledge [@Butler2018].  The scTree package was then used to then classify cells based on the top 5 markers for each cluster chosen by scTree and accurately recapitulates the original classification determined by Seurat.
 
 ## Example Output from the package
 
@@ -87,21 +87,16 @@ and connections with several antibody vendors are provided to query the availabi
 probes for the genes found to be usefull for classification.
 
 ``` r
-> as.garnett(clus6_ctree_fit, rules_keep = "^clus")
-# > cluster 6_node_11 	(n = 59)
-# expressed above: GNLY 3.479, GZMB 3.017
-#
-# > cluster 6_node_8 	(n = 14)
-# expressed above: FGFBP2 2.938
-# expressed below: GZMB 3.017
+as.garnett(tree_model, rules_keep = "^NK")
+# > NK Cells 	(n = 88)
+# expressed above: CST7 1.636, TYROBP 2.451
 ```
 
-**Code Section 1.** *Suggested classification scheme for cluster 6 of the PBMC dataset.*
-*The data depicts how Cluster 6 can be predominantly identified as GNLY High/GZMB High.*
-*nonetheless a minority can also be classified as GZMB Low/FGFBP2 High.*
+**Code Section 1.** *Suggested classification scheme for NK cell cluster of the PBMC dataset.*
+*The data depicts how the cluster corresponding to NK cells can be predominantly identified as GNLY High/GZMB High.*
 
-!["Scatterplot showing the progressive gating that would be used to classify node 11 in the 3K PBMC dataset. Filtering in each pane is donde on the gene presented on the X-axis of the plot and cells kept during that filtering step are highlighted."](./paper_figures/gating_clus_6.png "Figure 3")
[…]
