### 2020-03-16 #2061 @jspaezp
@natallah
Oddly enough, when previewing the paper using the `Whedon paper preview service`, the figure is correctly labelled.

here is the link to the current preview ...
https://whedon.theoj.org/preview?id=8b96fc6483534375133ccd09
