### 2020-03-16 #2061 @jenzopr
I can confirm that. The link sent by @jspaezp does not lead to the paper immediately, but if you use the service with the scTree repo URL as input, the figure labels are correct.
