### 2020-03-16 #2061 @arfon
@natallah - Pandoc knows how to number figures already so this change should fix your numbering issue: https://github.com/jspaezp/sctree/pull/13
