### 2020-03-21 202eeca
minor modification to the figure captions and merge details for figure labels
