### 2020-03-21 #2061 @jspaezp
hello @arfon,

I have accepted the figure label changes to master and updated all the active branches, Thanks for the PR. I also checked and they seem to render correctly using pandoc locally (`pandoc paper.md --filter pandoc-citeproc --bibliography bibliography.bib -o paper_$(echo $(date +%H%M)_$(date +%d%m%Y)).pdf`) .

Kindest wishes,
Sebastian

@natallah
