### 2020-03-23 7112e2e
minor typos in the paper and suggestions by the other authors
