### 2020-03-23 #2061 @jspaezp
I just pushed a commit that fixes a couple of minor typos and grammar details in the paper :D

https://github.com/jspaezp/sctree/commit/7112e2e5872fb241b3781024e3d33135379156b0
I feel like it should be good to go :D
@natallah
