### 2020-03-25 #2061 @mschubert
I apologize for the delay in updating my checklist, I was not feeling well the last 10 days (not sure if this virus or another).

Great work on this in the meantime @jspaezp! This all looks good to me now. I particularly like the more concise "mission statement" of the package.

<hr>

Here are some very minor, last discretionary changes that I do not consider acceptance blockers:
* Does JOSS have an opinion about snapshotting cited web pages to keep links accessible?
* Seurat tutorial (4th ref from bottom) does not provide a link
* There is a number of `R CMD check` warnings on Travis that could be fixed
