+++ vignettes/paper/bibliography.bib
@@ -203,5 +203,5 @@ year = {2019}
   title = {Seurat - Guided Clustering Tutorial},
   year = {2019},
   author = {"Satija-Lab"},
-  howoubished = {\url{https://satijalab.org/seurat/v3.1/pbmc3k_tutorial.html}}
+  howoubished = {\url{https://web.archive.org/web/20200309135737/https://satijalab.org/seurat/v3.1/pbmc3k_tutorial.html}}
 }
