+++ vignettes/paper/bibliography.bib
@@ -2,11 +2,9 @@
 @article{Janitza2018,
     author = {Janitza, Silke and Celik, Ender and Boulesteix, Anne-laure},
     doi = {10.1007/s11634-016-0276-4},
-    file = {:C$\backslash$:/Users/Sebastian/Downloads/s11634-016-0276-4.pdf:pdf},
     issn = {1862-5347},
     journal = {Advances in Data Analysis and Classification},
     keywords = {Feature selection,Gene selection,Random forests,Va},
-    mendeley-groups = {BioinfCore},
     month = {dec},
     number = {4},
     pages = {885--915},
@@ -19,10 +17,8 @@
 @article{Altmann2010,
     author = {Altmann, Andr{\'{e}} and Toloşi, Laura and Sander, Oliver and Lengauer, Thomas},
 doi = {10.1093/bioinformatics/btq134},
-file = {:C$\backslash$:/Users/Sebastian/Downloads/untitled.pdf:pdf},
 issn = {1460-2059},
 journal = {Bioinformatics},
-mendeley-groups = {BioinfCore},
 month = {may},
 number = {10},
 pages = {1340--1347},
@@ -34,11 +30,9 @@ year = {2010}
 @article{Wright2017,
 author = {Wright, Marvin N and Ziegler, Andreas},
 doi = {10.18637/jss.v077.i01},
-file = {:C$\backslash$:/Users/Sebastian/Downloads/v77i01.pdf:pdf},
 issn = {1548-7660},
 journal = {Journal of Statistical Software},
 keywords = {c,classification,machine learning,r,random forests,rcpp,recursive parti-,survival analysis,tioning},
-mendeley-groups = {BioinfCore},
 number = {1},
 title = {{ranger : A Fast Implementation of Random Forests for High Dimensional Data in C++ and R}},
 url = {http://www.jstatsoft.org/v77/i01/},
@@ -74,7 +68,6 @@ author = {Love, Michael I and Huber, Wolfgang and Anders, Simon},
 doi = {10.1186/s13059-014-0550-8},
 issn = {1474-760X},
 journal = {Genome biology},
-mendeley-groups = {Sequencing},
 number = {12},
 pages = {550},
 pmid = {25516281},
@@ -89,7 +82,6 @@ author = {Robinson, Mark D. and McCarthy, Davis J. and Smyth, Gordon K.},
 doi = {10.1093/bioinformatics/btp616},
 issn = {14602059},
 journal = {Bioinformatics},
-mendeley-groups = {Sequencing},
 number = {1},
 pages = {139--140},
 title = {{edgeR: A Bioconductor package for differential expression analysis of digital gene expression data}},
@@ -102,7 +94,6 @@ author = {Butler, Andrew and Hoffman, Paul and Smibert, Peter and Papalexi, Efth
 doi = {10.1038/nbt.4096},
 issn = {1546-1696},
 journal = {Nature biotechnology},
-mendeley-groups = {Sequencing},
 number = {5},
 pages = {411--420},
 pmid = {29608179},
@@ -117,7 +108,6 @@ author = {Trapnell, Cole and Cacchiarelli, Davide and Grimsby, Jonna and Pokhare
 doi = {10.1038/nbt.2859},
 issn = {1546-1696},
 journal = {Nature biotechnology},
-mendeley-groups = {Sequencing},
 number = {4},
 pages = {381--386},
 pmid = {24658644},
@@ -133,7 +123,6 @@ author = {Zheng, Grace X.Y. and Terry, Jessica M. and Belgrader, Phillip and Ryv
 doi = {10.1038/ncomms14049},
 issn = {20411723},
 journal = {Nature Communications},
[…]
