### 2020-04-02 #2061 @jspaezp
hello @mschubert I have modified the links in the bibliography.bib file and (hypothetically) fixed the key that made the link not show up in the references section.

regarding the check fixes ... for the most part its issues with the naming of arguments and patterns for s3 methods, which after trying would break A LOT of the checks and examples in the documentation, so I believe it should be done on changing to a major version. We would definitely add it to the milestones for v2.0. In addition the example data (another warn in the package check, due to the size) could be fixed by generating a data package independently and having it as a required dependency for automated testing but the warn would still appear in that one and make reasonably harder to set up the travis automated testing.

@natallah
