### 2020-04-19 #2061 @jspaezp
Hello @arfon ,

I was discussing with @natallah on what the best way to handle it would be.
On one hand I want it to be faithful to actual code output, but it going over the margins is not really useful information.
I am not familiar with latex, (alternative 1) is there a way to make the font on code chunks smaller for it to fit ?
We could also (alternative 2) just crop the content to fit the margins of the page and would resemble well enough the output of a terminal.

Highly appreciating your feedback on what would be the best approach,
Sebastian
