### 2020-04-19 #2061 @arfon
I think the second approach (cropping content) is probably the best one.
