### 2020-04-20 #2061 @jspaezp
Hello @arfon,

I cropped it in commit jspaezp/sctree@ac02e43abb49f9f8bf228b900e9f4178fa076273

@natallah
