### 2020-04-21 #2061 @arfon
@jspaezp - I made a couple of tiny formatting improvements in https://github.com/jspaezp/sctree/pull/14.

At this point could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
