### 2020-04-25 #2061 @arfon
@jenzopr - could you merge https://github.com/jspaezp/sctree/pull/14 please?
